requirements.txt
sqlmodel==0.0.22
SQLAlchemy==2.0.35
starlette # Не ограничивать версию пакета
pydantic==2.9.2
pydantic-settings==2.5.2
psycopg==3.2.2
psycopg-binary==3.2.2
email_validator==2.2.0
passlib==1.7.4
fastapi==0.115.0
loguru==0.7.2
PyJWT==2.9.0
redis==5.0.8
minio==7.2.8
alembic==1.13.2
uvicorn==0.30.6
python-multipart==0.0.9